Revamp make-connection to accomodate new authentication mechanism
The new authentication mechanism in Mongo driver 3.x requires
credentials to be passed to the MongoClient constructor.
This has the knock-on effect of requiring username/password to be passed
to make-connection.

Rather than trying to adapt the previous approach which pulled args out
of the fn args by type I've made the different optional connection args
explicit keyword-style arguments.

Bumped version to 0.6.0 because this is an incompatible API change.