Conditionally use clojure.core/seqable?
As of Clojure 1.9.0 Alpha 5, clojure.core has seqable? and it has been
deprecated in clojure.core.incubator. Since that was the only thing we
used incubator for, we now conditionally use clojure.core/seqable? if it
is available else we define an equivalent to incubator's seqable? from
the 1.3 version. This allows us to drop the incubator dependency.